Move to dep, move to circleci 2, extract mocks
