+++ .circleci/config.yml
@@ -1,0 +1,56 @@
+version: 2
+jobs:
+  build:
+    working_directory: /go/src/github.com/Financial-Times/smartlogic-notifier
+    docker:
+      - image: circleci/golang:1.10
+        environment:
+          CIRCLE_TEST_REPORTS: /tmp/test-results
+          CIRCLE_COVERAGE_REPORT: /tmp/coverage-results
+    steps:
+      - checkout
+      - run:
+          name: External dependencies
+          command: |
+            go get -u github.com/mattn/goveralls
+            go get -u github.com/jstemmer/go-junit-report
+            curl https://raw.githubusercontent.com/golang/dep/master/install.sh | sh
+      - run:
+          name: Make result folders
+          command: |
+            mkdir -p ${CIRCLE_TEST_REPORTS}
+            mkdir -p ${CIRCLE_COVERAGE_REPORT}
+      - run:
+          name: Download vendored dependancies
+          command: dep ensure -v -vendor-only
+      - run:
+          name: Go build
+          command: go build -v
+      - run:
+          name: Run tests
+          command: 
+            go test -race -cover -coverprofile=${CIRCLE_COVERAGE_REPORT}/coverage.out ./... | go-junit-report > ${CIRCLE_TEST_REPORTS}/junit.xml
+            sed -i "/_mock/d" ${CIRCLE_COVERAGE_REPORT}/coverage.out
+      - run:
+          name: Upload coverage
+          command: goveralls -coverprofile=${CIRCLE_COVERAGE_REPORT}/coverage.out -service=circle-ci -repotoken=${COVERALLS_TOKEN}
+      - store_test_results:
+          path: /tmp/test-results
+  dockerfile:
+    working_directory: /smartlogic-notifier
+    docker:
+      - image: docker:18.03
+    steps:
+      - checkout
+      - setup_docker_engine
+      - run:
+          name: Build Dockerfile
+          command: docker build .
+workflows:
+  version: 2
+  test-and-build-docker:
+    jobs:
+      - build
+      - dockerfile:
+          requires:
+            - build
