+++ Gopkg.lock
@@ -1,0 +1,181 @@
+# This file is autogenerated, do not edit; changes may be undone by the next 'dep ensure'.
+
+
+[[projects]]
+  name = "github.com/Financial-Times/go-fthealth"
+  packages = ["v1_1"]
+  revision = "1b007e2b37b7936dfb6671fa17e5a29fd35a08ea"
+  version = "0.4.0"
+
+[[projects]]
+  name = "github.com/Financial-Times/http-handlers-go"
+  packages = ["httphandlers"]
+  revision = "229ac16f1d9ec9bca485d2dafb9ec14dc7e9ba24"
+  version = "0.1.1"
+
+[[projects]]
+  name = "github.com/Financial-Times/kafka-client-go"
+  packages = ["kafka"]
+  revision = "cc65f68255d42de893a6d72f31394ad031c54142"
+  version = "0.4.0"
+
+[[projects]]
+  name = "github.com/Financial-Times/service-status-go"
+  packages = [
+    "buildinfo",
+    "gtg",
+    "httphandlers"
+  ]
+  revision = "3f5199736a3d7ae52394c63aac36834786825e21"
+  version = "0.1.0"
+
+[[projects]]
+  name = "github.com/Financial-Times/transactionid-utils-go"
+  packages = ["."]
+  revision = "df2f00c734957c9dd651ce23ab0e0902504c7636"
+  version = "v0.2.0"
+
+[[projects]]
+  name = "github.com/Shopify/sarama"
+  packages = ["."]
+  revision = "ac03dfa0c99a940a48357980a4ebfafed5f23f9a"
+
+[[projects]]
+  name = "github.com/davecgh/go-spew"
+  packages = ["spew"]
+  revision = "346938d642f2ec3594ed81d874461961cd0faa76"
+  version = "v1.1.0"
+
+[[projects]]
+  name = "github.com/eapache/go-resiliency"
+  packages = ["breaker"]
+  revision = "b1fe83b5b03f624450823b751b662259ffc6af70"
+
+[[projects]]
+  branch = "master"
+  name = "github.com/eapache/go-xerial-snappy"
+  packages = ["."]
+  revision = "bb955e01b9346ac19dc29eb16586c90ded99a98c"
+
+[[projects]]
+  name = "github.com/eapache/queue"
+  packages = ["."]
+  revision = "44cc805cf13205b55f69e14bcb69867d1ae92f98"
+  version = "v1.1.0"
+
+[[projects]]
+  branch = "master"
+  name = "github.com/golang/snappy"
+  packages = ["."]
+  revision = "553a641470496b2327abcac10b36396bd98e45c9"
+
+[[projects]]
+  branch = "master"
+  name = "github.com/gorilla/context"
+  packages = ["."]
+  revision = "08b5f424b9271eedf6f9f0ce86cb9396ed337a42"
+
+[[projects]]
+  name = "github.com/gorilla/handlers"
+  packages = ["."]
+  revision = "a4043c62cc2329bacda331d33fc908ab11ef0ec3"
+  version = "v1.2.1"
+
+[[projects]]
+  name = "github.com/gorilla/mux"
+  packages = ["."]
+  revision = "043ee6597c29786140136a5747b6a886364f5282"
+
+[[projects]]
+  name = "github.com/hashicorp/go-version"
+  packages = ["."]
+  revision = "03c5bf6be031b6dd45afec16b1cf94fc8938bc77"
+
+[[projects]]
+  name = "github.com/jawher/mow.cli"
+  packages = ["."]
+  revision = "8327d12beb75e6471b7f045588acc318d1147146"
+
+[[projects]]
+  name = "github.com/joho/godotenv"
+  packages = [
+    ".",
+    "autoload"
+  ]
+  revision = "325433c502d409f3c3dc820098fb0cfe38d98dc7"
+
+[[projects]]
+  name = "github.com/pierrec/lz4"
+  packages = ["."]
+  revision = "5a3d2245f97fc249850e7802e3c01fad02a1c316"
+
+[[projects]]
+  name = "github.com/pierrec/xxHash"
+  packages = ["xxHash32"]
+  revision = "5a004441f897722c627870a981d02b29924215fa"
+
+[[projects]]
+  name = "github.com/pmezard/go-difflib"
+  packages = ["difflib"]
+  revision = "792786c7400a136282c1664665ae0a8db921c6c2"
+  version = "v1.0.0"
+
+[[projects]]
+  name = "github.com/rcrowley/go-metrics"
+  packages = ["."]
+  revision = "1f30fe9094a513ce4c700b9a54458bbb0c96996c"
+
+[[projects]]
+  name = "github.com/samuel/go-zookeeper"
+  packages = ["zk"]
+  revision = "1d7be4effb13d2d908342d349d71a284a7542693"
+
+[[projects]]
+  name = "github.com/sethgrid/pester"
+  packages = ["."]
+  revision = "4f4c0a67b6496764028e1ab9fd8dfb630282ed2f"
+
+[[projects]]
+  name = "github.com/sirupsen/logrus"
+  packages = ["."]
+  revision = "89742aefa4b206dcf400792f3bd35b542998eb3b"
+
+[[projects]]
+  name = "github.com/stretchr/testify"
+  packages = ["assert"]
+  revision = "f6abca593680b2315d2075e0f5e2a9751e3f431a"
+
+[[projects]]
+  name = "github.com/wvanbergen/kafka"
+  packages = ["consumergroup"]
+  revision = "bc265fedb9ff5b5c5d3c0fdcef4a819b3523d3ee"
+
+[[projects]]
+  name = "github.com/wvanbergen/kazoo-go"
+  packages = ["."]
+  revision = "968957352185472eacb69215fa3dbfcfdbac1096"
+
+[[projects]]
+  name = "golang.org/x/crypto"
+  packages = ["ssh/terminal"]
+  revision = "eb71ad9bd329b5ac0fd0148dd99bd62e8be8e035"
+
+[[projects]]
+  name = "golang.org/x/net"
+  packages = ["context"]
+  revision = "3da985ce5951d99de868be4385f21ea6c2b22f24"
+
+[[projects]]
+  name = "golang.org/x/sys"
+  packages = [
+    "unix",
+    "windows"
+  ]
+  revision = "07c182904dbd53199946ba614a412c61d3c548f5"
+
+[solve-meta]
+  analyzer-name = "dep"
+  analyzer-version = 1
+  inputs-digest = "5056af9254cbc41e0dd039fba676fe849fc886f9304379911c634c6374f64c9a"
+  solver-name = "gps-cdcl"
+  solver-version = 1
