+++ Gopkg.toml
@@ -1,0 +1,54 @@
+# Gopkg.toml example
+#
+# Refer to https://github.com/golang/dep/blob/master/docs/Gopkg.toml.md
+# for detailed Gopkg.toml documentation.
+#
+# required = ["github.com/user/thing/cmd/thing"]
+# ignored = ["github.com/user/project/pkgX", "bitbucket.org/user/project/pkgA/pkgY"]
+#
+# [[constraint]]
+#   name = "github.com/user/project"
+#   version = "1.0.0"
+#
+# [[constraint]]
+#   name = "github.com/user/project2"
+#   branch = "dev"
+#   source = "github.com/myfork/project2"
+#
+# [[override]]
+#   name = "github.com/x/y"
+#   version = "2.4.0"
+#
+# [prune]
+#   non-go = false
+#   go-tests = true
+#   unused-packages = true
+
+
+[[constraint]]
+  name = "github.com/Financial-Times/go-fthealth"
+  version = "0.4.0"
+
+[[constraint]]
+  name = "github.com/Financial-Times/http-handlers-go"
+  version = "0.1.1"
+
+[[constraint]]
+  name = "github.com/Financial-Times/kafka-client-go"
+  version = "0.4.0"
+
+[[constraint]]
+  name = "github.com/Financial-Times/service-status-go"
+  version = "0.1.0"
+
+[[constraint]]
+  name = "github.com/Financial-Times/transactionid-utils-go"
+  version = "0.2.0"
+
+[[constraint]]
+  name = "github.com/gorilla/handlers"
+  version = "1.2.1"
+
+[prune]
+  go-tests = true
+  unused-packages = true
