--- circle.yml
@@ -1,30 +1,0 @@
-machine:
-  environment:
-    PROJECT_GOPATH: "${HOME}/.go_project"
-    PROJECT_PARENT_PATH: "${PROJECT_GOPATH}/src/github.com/${CIRCLE_PROJECT_USERNAME}"
-    PROJECT_PATH: "${PROJECT_PARENT_PATH}/${CIRCLE_PROJECT_REPONAME}"
-    GOPATH: "${HOME}/.go_workspace:${PROJECT_GOPATH}"
-
-checkout:
-  post:
-    - mkdir -p "${PROJECT_PARENT_PATH}"
-    - ln -sf "${HOME}/${CIRCLE_PROJECT_REPONAME}/" "${PROJECT_PATH}"
-
-dependencies:
-  pre:
-    - go get -u github.com/kardianos/govendor
-  override:
-    - cd $PROJECT_PATH && govendor sync
-    - cd $PROJECT_PATH && go build -v
-
-test:
-  pre:
-    - go get -u github.com/jstemmer/go-junit-report
-    - go get -u github.com/mattn/goveralls
-    - cd $PROJECT_PATH && wget https://raw.githubusercontent.com/Financial-Times/cookiecutter-upp-golang/master/coverage.sh && chmod +x coverage.sh
-  override:
-    - mkdir -p $CIRCLE_TEST_REPORTS/golang
-    - cd $PROJECT_PATH && govendor test -race -v +local | go-junit-report > $CIRCLE_TEST_REPORTS/golang/junit.xml
-    - cd $PROJECT_PATH && ./coverage.sh
-  post:
-    - goveralls -coverprofile=$CIRCLE_ARTIFACTS/coverage.out -service=circle-ci -repotoken=$COVERALLS_TOKEN
